+++ kelvin-api/ucsschool/kelvin/routers/user.py
@@ -87,7 +87,7 @@ from univention.admin.filter import conjunction, expression
 from ..config import UDM_MAPPING_CONFIG
 from ..import_config import get_import_config, init_ucs_school_import_framework
 from ..ldap import LdapUser, get_dn_of_user
-from ..token_auth import get_kelvin_admin
+from ..token_auth import get_kelvin_admin, get_kelvin_reader
 from ..urls import cached_url_for, url_to_name
 from .base import (
     APIAttributesMixin,
@@ -606,7 +606,7 @@ async def search(  # noqa: C901
     logger: logging.Logger = Depends(get_logger),
     accepted_properties: Set[str] = Depends(accepted_udm_properties),
     udm: UDM = Depends(udm_ctx),
-    kelvin_admin: LdapUser = Depends(get_kelvin_admin),
+    authenticated_user: LdapUser = Depends(get_kelvin_reader),
 ) -> List[UserModel]:
     """
     Search for school users.
@@ -744,7 +744,7 @@ async def get(
     username: str = Path(default=..., description="Name of the school user to fetch."),
     logger: logging.Logger = Depends(get_logger),
     udm: UDM = Depends(udm_ctx),
-    kelvin_admin: LdapUser = Depends(get_kelvin_admin),
+    authenticated_user: LdapUser = Depends(get_kelvin_reader),
 ) -> UserModel:
     """
     Fetch a specific school user.
@@ -771,7 +771,7 @@ async def create(
     ),
     logger: logging.Logger = Depends(get_logger),
     udm: UDM = Depends(udm_ctx),
-    kelvin_admin: LdapUser = Depends(get_kelvin_admin),
+    authenticated_user: LdapUser = Depends(get_kelvin_admin),
 ) -> UserModel:
     """
     Create a school user with all the information:
@@ -1035,7 +1035,7 @@ async def partial_update(  # noqa: C901
     request: Request,
     logger: logging.Logger = Depends(get_logger),
     udm: UDM = Depends(udm_ctx),
-    kelvin_admin: LdapUser = Depends(get_kelvin_admin),
+    authenticated_user: LdapUser = Depends(get_kelvin_admin),
 ) -> UserModel:
     """
     Patch a school **user** with partial information
@@ -1224,7 +1224,7 @@ async def complete_update(  # noqa: C901
     request: Request,
     logger: logging.Logger = Depends(get_logger),
     udm: UDM = Depends(udm_ctx),
-    kelvin_admin: LdapUser = Depends(get_kelvin_admin),
+    authenticated_user: LdapUser = Depends(get_kelvin_admin),
 ) -> UserModel:
     """
     Update a school **user** with all the information:
@@ -1420,7 +1420,7 @@ async def delete(
     username: str,
     request: Request,
     udm: UDM = Depends(udm_ctx),
-    kelvin_admin: LdapUser = Depends(get_kelvin_admin),
+    authenticated_user: LdapUser = Depends(get_kelvin_admin),
 ) -> Response:
     """
     Delete a school user