+++ kelvin-api/ucsschool/kelvin/routers/workgroup.py
@@ -39,7 +39,7 @@ from udm_rest_client import UDM, CreateError, ModifyError
 
 from ..config import UDM_MAPPING_CONFIG
 from ..ldap import LdapUser
-from ..token_auth import get_kelvin_admin
+from ..token_auth import get_kelvin_admin, get_kelvin_reader
 from ..urls import cached_url_for, url_to_dn, url_to_name
 from .base import (
     APIAttributesMixin,
@@ -181,7 +181,7 @@ async def search(
         title="name",
     ),
     udm: UDM = Depends(udm_ctx),
-    kelvin_admin: LdapUser = Depends(get_kelvin_admin),
+    authenticated_user: LdapUser = Depends(get_kelvin_reader),
 ) -> List[WorkGroupModel]:
     """
     Search for school workgroups.
@@ -205,7 +205,7 @@ async def get(
     school: str,
     request: Request,
     udm: UDM = Depends(udm_ctx),
-    kelvin_admin: LdapUser = Depends(get_kelvin_admin),
+    authenticated_user: LdapUser = Depends(get_kelvin_reader),
 ) -> WorkGroupModel:
     sc = await get_lib_obj(udm, WorkGroup, f"{school}-{workgroup_name}", school)
     return await WorkGroupModel.from_lib_model(sc, request, udm)
@@ -217,7 +217,7 @@ async def create(
     request: Request,
     udm: UDM = Depends(udm_ctx),
     logger: logging.Logger = Depends(get_logger),
-    kelvin_admin: LdapUser = Depends(get_kelvin_admin),
+    authenticated_user: LdapUser = Depends(get_kelvin_admin),
 ) -> WorkGroupModel:
     """
     Create a **workgroup** with all the information:
@@ -292,7 +292,7 @@ async def partial_update(
     request: Request,
     udm: UDM = Depends(udm_ctx),
     logger: logging.Logger = Depends(get_logger),
-    kelvin_admin: LdapUser = Depends(get_kelvin_admin),
+    authenticated_user: LdapUser = Depends(get_kelvin_admin),
 ) -> WorkGroupModel:
     """
     Update a **workgroup** with all the information:
@@ -363,7 +363,7 @@ async def complete_update(
     request: Request,
     udm: UDM = Depends(udm_ctx),
     logger: logging.Logger = Depends(get_logger),
-    kelvin_admin: LdapUser = Depends(get_kelvin_admin),
+    authenticated_user: LdapUser = Depends(get_kelvin_admin),
 ) -> WorkGroupModel:
     """
     Update a **workgroup** with all the information:
@@ -450,7 +450,7 @@ async def delete(
     school: str,
     request: Request,
     udm: UDM = Depends(udm_ctx),
-    kelvin_admin: LdapUser = Depends(get_kelvin_admin),
+    authenticated_user: LdapUser = Depends(get_kelvin_admin),
 ) -> Response:
     sc = await get_lib_obj(udm, WorkGroup, f"{school}-{workgroup_name}", school)
     await sc.remove(udm)